Error patching, courtesy of the new Shields Up supervisor
Should I incorporate your plugin into the Shields Up repository and then give you access to edit it? That would certainly make it easier to have everything in one spot, and when you want to release an independent version then just copy out the Alphean part of the mod pack.